Show a warning if the trip is entirely within the bubble of humanity.
There are many historical/civilisation POIs near Earth because it is
the centre of human existence in the galaxy, and so we ignore them from
scenic route calculations. (Otherwise it would be like saying you are
going from London to New York, so on your way you should visit Big Ben,
the British Museum, etc.)

People may not understand this and be confused that there are no results,
so we now show a polite error instead.